refactor(web): split spider_chrome into sub-mods
Organize spider_chrome backend into focused modules with backend core,
error types, CLI args, and utility functions. Maintain same public API
while improving code organization and maintainability.
